AddEditLead changings Init Service finalization Response of InitService in NavigationBottomMain is currently commented due to any potential instability and initial version in production.

+++ app/src/main/java/com/example/muzafarimran/lastingsales/service/MyFirebaseMessagingService.java
@@ -220,11 +220,13 @@ public class MyFirebaseMessagingService extends FirebaseMessagingService {
                 } else if (action.equals("delete")) {
                     String id = payload.getString("id");
                     LSContact contact = LSContact.getContactFromServerId(id);
-                    Log.e(TAG, "handleDataMessage: contact: " + contact.toString());
-                    contact.delete();
-                    LeadContactAddedEventModel mCallEvent = new LeadContactAddedEventModel();
-                    TinyBus bus = TinyBus.from(getApplicationContext());
-                    bus.post(mCallEvent);
+                    if (contact != null) {
+                        Log.e(TAG, "handleDataMessage: contact: " + contact.toString());
+                        contact.delete();
+                        LeadContactAddedEventModel mCallEvent = new LeadContactAddedEventModel();
+                        TinyBus bus = TinyBus.from(getApplicationContext());
+                        bus.post(mCallEvent);
+                    }
                 }
             }
